Fix warnings, make logging stuff consistent with doobie.
