+++ doobie-support/src/main/scala/scoobie/doobie/QueryExpressionExtensions.scala
@@ -9,6 +9,5 @@ import scoobie.ast.QueryExpression
   * Created by jacob.barber on 5/25/16.
   */
 class QueryExpressionExtensions(expr: QueryExpression[ScoobieFragmentProducer])(implicit sqlInterpreter: SqlInterpreter) {
-  def buildAndPrint(printer: Fragment => Unit): Fragment = sqlInterpreter.genAndPrintSql(expr, printer)
-  def build: Fragment = buildAndPrint(void)
+  def build: Fragment = sqlInterpreter.genSql(expr)
 }
