+++ doobie-support/src/main/scala/scoobie/doobie/QueryModifyExtensions.scala
@@ -8,6 +8,6 @@ import scoobie.ast.QueryModify
   * Created by jacob.barber on 5/25/16.
   */
 class QueryModifyExtensions(expr: QueryModify[ScoobieFragmentProducer])(implicit sqlInterpreter: SqlInterpreter) {
-  def updateAndPrint(printer: Fragment => Unit): Update0 = sqlInterpreter.update(expr, printer)
-  def update: Update0 = updateAndPrint(void)
+  def updateWithLogHandler(logHandler: LogHandler): Update0 = sqlInterpreter.update(expr, logHandler)
+  def update(implicit logHandler: LogHandler = LogHandler.nop): Update0 = sqlInterpreter.update(expr, logHandler)
 }
