+++ doobie-support/src/main/scala/scoobie/doobie/SqlInterpreter.scala
@@ -13,15 +13,9 @@ import shapeless.HList
   * Created by jacob.barber on 5/25/16.
   */
 case class SqlInterpreter(genSql: QueryExpression[ScoobieFragmentProducer] => Fragment) {
-  def query[B: Composite](ast: QuerySelect[ScoobieFragmentProducer], printer: Fragment => Unit): Query0[B] =
-    genAndPrintSql(ast, printer).queryWithLogHandler[B](LogHandler.jdkLogHandler)
+  def query[B: Composite](ast: QuerySelect[ScoobieFragmentProducer], logHandler: LogHandler): Query0[B] =
+    genSql(ast).queryWithLogHandler(logHandler)
 
-  def update(ast: QueryModify[ScoobieFragmentProducer], printer: Fragment => Unit): Update0 =
-    genAndPrintSql(ast, printer).updateWithLogHandler(LogHandler.jdkLogHandler)
-
-  def genAndPrintSql(ast: QueryExpression[ScoobieFragmentProducer], printer: Fragment => Unit): Fragment = {
-    val sqlString = genSql(ast)
-    printer(sqlString)
-    sqlString
-  }
+  def update(ast: QueryModify[ScoobieFragmentProducer], logHandler: LogHandler): Update0 =
+    genSql(ast).updateWithLogHandler(logHandler)
 }
